MdeModulePkg: Set Page Writeable Before Attempting to Debug Clear Memory
Currently, if the DebugClearMemory bit is set in the
PcdDebugPropertyMask, CoreConvertPagesEx will attempt to write a pattern
to the pages being freed. However, it does not check that the page is
writeable, which will cause a page fault if not.

This patch updates the page freeing code to remove EFI_MEMORY_RO and
EFI_MEMORY_RP from pages if we are doing to do debug clear memory on
them. It does this through the GCD to keep it in sync with the
page table state. Pool Pages do not have this applied because the GCD
allocates and frees pool memory on every operation, which could cause
a recursive call into the GCD, tripping the lock. Page attributes
should never be set on pool memory (except by the core), because it
affects the entire page, which may have different owners of other
pools, so we expect pool pages to always be writeable.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>