EmbeddedPkg: PrePiMemoryAllocationLib: Add AllocateCopyPool
ArmVirtPkg uses PrePiMemoryAllocationLib and PeiServicesLib in
the same module. PeiServicesLib depends on MemoryAllocationLib,
fulfilled by PrePiMemoryAllocationLib in this instance. However,
PeiServicesLib uses AllocateCopyPool() which is not provided in
PrePiMemoryAllocationLib. CLANGPDB AARCH64 finds this as a link
error. Other toolchains are optimizing the call out since that
code path is not used in this module.

This adds AllocateCopyPool() to PrePiMemoryAllocationLib to
satisfy the dependency.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>